Refactor ModulesUtil to use ScheduleUtil and MonthDomainValues
- Deprecate legacy enums and methods.
